+++ src/wireless/thread.rs
@@ -22,15 +22,14 @@ use crate::mdns::Mdns;
 use crate::nal::NetStack;
 use crate::network::Embedding;
 use crate::wireless::{GattPeripheral, GattTask, MatterStackWirelessTask, WirelessNetCtl};
-use crate::{pin_alloc, UserTask};
+use crate::UserTask;
 
 use super::{Gatt, PreexistingWireless, WirelessMatterStack};
 
 /// A type alias for a Matter stack running over Thread (and BLE, during commissioning).
-pub type ThreadMatterStack<'a, const B: usize, E = ()> =
-    WirelessMatterStack<'a, B, wireless::Thread, E>;
+pub type ThreadMatterStack<'a, E = ()> = WirelessMatterStack<'a, wireless::Thread, E>;
 
-impl<const B: usize, E> WirelessMatterStack<'_, B, wireless::Thread, E>
+impl<E> WirelessMatterStack<'_, wireless::Thread, E>
 where
     E: Embedding,
 {
@@ -106,18 +105,9 @@ where
 
         info!("Matter Stack memory: {}b", core::mem::size_of_val(self));
 
-        // Since this is the last code executed in the method, resetting the allocator should be safe
-        // because all boxes returned by it should be dropped by then
-        let _defer = scopeguard::guard((), |_| unsafe {
-            self.bump.reset();
-        });
-
         self.matter().reset_transport()?;
 
-        let net_task = pin_alloc!(
-            self.bump,
-            self.run_thread_coex(&mut thread, crypto, handler, kv, user)
-        );
+        let net_task = pin!(self.run_thread_coex(&mut thread, crypto, handler, kv, user));
 
         net_task.await
     }
@@ -149,18 +139,9 @@ where
 
         info!("Matter Stack memory: {}b", core::mem::size_of_val(self));
 
-        // Since this is the last code executed in the method, resetting the allocator should be safe
-        // because all boxes returned by it should be dropped by then
-        let _defer = scopeguard::guard((), |_| unsafe {
-            self.bump.reset();
-        });
-
         self.matter().reset_transport()?;
 
-        let net_task = pin_alloc!(
-            self.bump,
-            self.run_thread(thread, crypto, handler, kv, user)
-        );
+        let net_task = pin!(self.run_thread(thread, crypto, handler, kv, user));
 
         net_task.await
     }
@@ -185,7 +166,7 @@ where
         // `&kv` is also lent to the driver so it can persist its own state.
         thread
             .run(
-                MatterStackWirelessTask::<'_, _, _, _, _, _, _, _, NoopWirelessNetCtl> {
+                MatterStackWirelessTask::<'_, _, _, _, _, _, _, NoopWirelessNetCtl> {
                     stack: self,
                     crypto,
                     handler,
@@ -227,7 +208,7 @@ where
 
                 Gatt::run(
                     &mut thread,
-                    MatterStackWirelessTask::<'_, _, _, _, _, _, _, _, <W as Thread>::NetCtl<'_>> {
+                    MatterStackWirelessTask::<'_, _, _, _, _, _, _, <W as Thread>::NetCtl<'_>> {
                         stack: self,
                         crypto: &crypto,
                         handler: &handler,
@@ -241,7 +222,7 @@ where
 
             Thread::run(
                 &mut thread,
-                MatterStackWirelessTask::<'_, _, _, _, _, _, _, _, <W as Thread>::NetCtl<'_>> {
+                MatterStackWirelessTask::<'_, _, _, _, _, _, _, <W as Thread>::NetCtl<'_>> {
                     stack: self,
                     crypto: &crypto,
                     handler: &handler,
@@ -453,8 +434,8 @@ where
     }
 }
 
-impl<'a, const B: usize, E, C, H, K, X, Q> GattTask
-    for MatterStackWirelessTask<'a, B, wireless::Thread, E, C, H, K, X, Q>
+impl<'a, E, C, H, K, X, Q> GattTask
+    for MatterStackWirelessTask<'a, wireless::Thread, E, C, H, K, X, Q>
 where
     E: Embedding,
     C: Crypto,
@@ -499,8 +480,8 @@ where
     }
 }
 
-impl<'a, const B: usize, E, C, H, K, X, Z> ThreadTask
-    for MatterStackWirelessTask<'a, B, wireless::Thread, E, C, H, K, X, Z>
+impl<'a, E, C, H, K, X, Z> ThreadTask
+    for MatterStackWirelessTask<'a, wireless::Thread, E, C, H, K, X, Z>
 where
     E: Embedding,
     C: Crypto,
@@ -615,8 +596,8 @@ where
     }
 }
 
-impl<'a, const B: usize, E, C, H, K, X, Z> ThreadCoexTask
-    for MatterStackWirelessTask<'a, B, wireless::Thread, E, C, H, K, X, Z>
+impl<'a, E, C, H, K, X, Z> ThreadCoexTask
+    for MatterStackWirelessTask<'a, wireless::Thread, E, C, H, K, X, Z>
 where
     E: Embedding,
     C: Crypto,
@@ -667,16 +648,12 @@ where
         );
 
         let stack = &self.stack;
-        let bump = &stack.bump;
+        let mut net_task =
+            pin!(stack.run_net_coex(&self.crypto, &net_stack, &netif, &mut mdns, &mut gatt));
 
-        let mut net_task = pin_alloc!(
-            bump,
-            stack.run_net_coex(&self.crypto, &net_stack, &netif, &mut mdns, &mut gatt)
-        );
+        let mut im_task = pin!(self.stack.run_im(&im));
 
-        let mut im_task = pin_alloc!(bump, self.stack.run_im_with_bump(&im));
-
-        let mut user_task = pin_alloc!(bump, self.user_task.run(&net_stack, &netif));
+        let mut user_task = pin!(self.user_task.run(&net_stack, &netif));
 
         select3(&mut net_task, &mut im_task, &mut user_task)
             .coalesce()